backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Sean Daly", email = "the_sean_c@fastmail.com"},
]
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn>=0.32.1",
    "sqlalchemy>=2.0.36",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.17",
    "pydantic[email]>=2.10.2",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
    "asyncpg>=0.29.0",
    "alembic>=1.14.0",
    "greenlet>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "httpx>=0.27.2",
    "click>=8.1.7",
    "cachetools>=5.5.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
package-dir = "src"
source-includes = ["tests/", "alembic/"]

[tool.pdm.scripts]
start = "uvicorn backend.main:app --reload"
migrate = "alembic upgrade head"
test = "pytest tests/"
